adding new samples
