+++ genai/image-generation/imggen-mmflash-locale-aware-with-txt.js
@@ -21,7 +21,6 @@ const {GoogleGenAI, Modality} = require('@google/genai');
 const GOOGLE_CLOUD_PROJECT = process.env.GOOGLE_CLOUD_PROJECT;
 const GOOGLE_CLOUD_LOCATION =
   process.env.GOOGLE_CLOUD_LOCATION || 'us-central1';
-const FILE_NAME = 'test/test-data/example-image-eiffel-tower.png';
 
 async function generateImage(
   projectId = GOOGLE_CLOUD_PROJECT,
@@ -33,11 +32,9 @@ async function generateImage(
     location: location,
   });
 
-  const image = fs.readFileSync(FILE_NAME);
-
   const response = await client.models.generateContent({
     model: 'gemini-2.5-flash-image-preview',
-    contents: [image, 'Edit this image to make it look like a cartoon'],
+    contents: 'Generate a photo of a breakfast meal.',
     config: {
       responseModalities: [Modality.TEXT, Modality.IMAGE],
     },
@@ -49,15 +46,22 @@ async function generateImage(
     if (part.text) {
       console.log(`${part.text}`);
     } else if (part.inlineData) {
+      const outputDir = 'output-folder';
+      if (!fs.existsSync(outputDir)) {
+        fs.mkdirSync(outputDir, { recursive: true });
+      }
       const imageBytes = Buffer.from(part.inlineData.data, 'base64');
-      fs.writeFileSync('bw-example-image.png', imageBytes);
+      const filename = `${outputDir}/example-breakfast-meal.png`;
+      fs.writeFileSync(filename, imageBytes);
     }
   }
 
   return response;
 }
 // Example response:
-// Okay, I will edit this image to give it a cartoonish style, with bolder outlines, simplified details, and more vibrant colors.
+// Generates a photo of a vibrant and appetizing breakfast meal.
+// The scene will feature a white plate with golden-brown pancakes
+// stacked neatly, drizzled with rich maple syrup and ...
 // [END googlegenaisdk_imggen_mmflash_edit_img_with_txt_img]
 
 module.exports = {
