+++ genai/package.json
@@ -16,6 +16,7 @@
     "@google/genai": "1.20.0",
     "axios": "^1.6.2",
     "luxon": "^3.7.1",
+    "proxyquire": "^2.1.3",
     "supertest": "^7.0.0"
   },
   "devDependencies": {
