+++ genai/test-data/latte.jpg
