+++ genai/test-data/scones.jpg
