+++ genai/test-data/sweater.jpg
