+++ genai/test/imggen-mmflash-locale-aware-with-txt.test.js
@@ -18,18 +18,16 @@ const {assert} = require('chai');
 const {describe, it} = require('mocha');
 
 const projectId = process.env.CAIP_PROJECT_ID;
-// const sample = require('../image-generation/imggen_virtual-try-on-with-txt-img');
-// const {delay} = require('./util');
+const location = 'global';
+const sample = require('../image-generation/imggen-mmflash-locale-aware-with-txt');
+const {delay} = require('./util');
 
-const sample = require('../september-samples/imggen_virtual-try-on-with-txt-img');
-const {delay} = require('../test/util');
-
-describe('imggen_virtual-try-on-with-txt-img', async () => {
-  it('should return a response object containing image parts', async function () {
+describe('imggen-mmflash-locale-aware-with-txt', async () => {
+  it('should generate a response with text and image parts', async function () {
     this.timeout(180000);
-    // this.retries(4);
-    // await delay(this.test);
-    const response = await sample.virtualTryOn(projectId);
+    this.retries(4);
+    await delay(this.test);
+    const response = await sample.generateImage(projectId, location);
     assert(response);
   });
 });
