+++ genai/test/imggen-mmflash-multiple-imgs-with-txt.test.js
@@ -18,15 +18,16 @@ const {assert} = require('chai');
 const {describe, it} = require('mocha');
 
 const projectId = process.env.CAIP_PROJECT_ID;
-const sample = require('../image-generation/imggen-mmflash-txt-and-img-with-txt');
+
+const sample = require('../image-generation/imggen-mmflash-multiple-imgs-with-txt');
 const {delay} = require('./util');
 
-describe('imggen-mmflash-txt-and-img-with-txt', async () => {
-  it('should generate a response with text and image parts', async function () {
+describe('imggen-mmflash-multiple-imgs-with-txt', async () => {
+  it('should return a response object containing image parts', async function () {
     this.timeout(180000);
     this.retries(4);
     await delay(this.test);
     const response = await sample.generateImage(projectId);
-    assert(response.length > 0);
+    assert(response);
   });
 });
