+++ genai/test/imggen-mmflash-with-txt.test.js
@@ -24,10 +24,9 @@ const {delay} = require('./util');
 describe('imggen-mmflash-with-txt', async () => {
   it('should generate images from a text prompt', async function () {
     this.timeout(180000);
-    this.retries(10);
+    this.retries(4);
     await delay(this.test);
-    const generatedFileNames = await sample.generateContent(projectId);
-    assert(Array.isArray(generatedFileNames));
+    const generatedFileNames = await sample.generateImage(projectId);
     assert(generatedFileNames.length > 0);
   });
 });
