+++ genai/test/imggen_virtual-try-on-with-txt-img.test.js
@@ -18,18 +18,16 @@ const {assert} = require('chai');
 const {describe, it} = require('mocha');
 
 const projectId = process.env.CAIP_PROJECT_ID;
-// const sample = require('../image-generation/imggen_virtual-try-on-with-txt-img');
-// const {delay} = require('./util');
 
-const sample = require('../september-samples/imggen-mmflash-multiple-imgs-with-txt');
-const {delay} = require('../test/util');
+const sample = require('../image-generation/imggen_virtual-try-on-with-txt-img');
+const {delay} = require('./util');
 
-describe('imggen-mmflash-multiple-imgs-with-txt', async () => {
+describe('imggen_virtual-try-on-with-txt-img', async () => {
   it('should return a response object containing image parts', async function () {
     this.timeout(180000);
-    // this.retries(4);
-    // await delay(this.test);
-    const response = await sample.generateImage(projectId);
+    this.retries(4);
+    await delay(this.test);
+    const response = await sample.virtualTryOn(projectId);
     assert(response);
   });
 });
